requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
distro==1.9.0
dotenv==0.9.9
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
jiter==0.9.0
openai==1.70.0
pydantic==2.11.1
pydantic_core==2.33.0
python-dotenv==1.1.0
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
